Add comments. Keeping manual detector creation.
